This is the DEFINITIONS revision. Here, we added a new section, the Definitions section, in order to define G1 and G2, and thereby to control Lule Sámi dipth simplification (ok in G12, not ok in G3).
At the end of the day, we were not able to check for just that, but at
least we got the definition section up and working (moral: here you
can use set-theoretic operators), and we corrected some errors in the
oa:å dipht simpl rule. We also experimented with removing Part I of
this rule, and declaring the first o:0 of part I as part of part
II. When doing so, we got an error message ("o-stem vowel deletion in
short passives" fails in state 3.) for the test case boalloX4j quoted
below, but it might well be that this error is secondary, and that the
main approach to Part I / Part II merger is sound. TODO: Check whether
it is so.