Lots of changes! ================
Thomas, Trond and Sjur had a long session, mainly to correct (lack of) diphtong simplification in "invisible" G3 stems. During the course of action, other errors were corrected as well, and several more test pairs added.

In addition to the substantial changes, very many rules were changed graphically, in order to improve their readability, and thus ease debugging of the twol rules

Definition changes:
===================

We corrected a typo in our G2/G3 definitions, earlier the exceptional three-letter G2 had the sequence " s k " twice, now the first of them was corrected to " s t ". It was simply a typo in the previous version.

We inserted Null symbols in our G12, G2 definitions. Earlier, the grade definitions worked only for cases where the wordform in question had undergone no changes, by adding zeros (in the form (Cns:0)") in appropriate places, we made the definitions usable also in cases where the forms actually underwent deletions.

The G123 definitions were changed to "LowerG2, LowerG2" yesterday. We made the definitions in order to block diphthong simplification in G3 cases, and one problem turned out to be that we in strengthening cases G2/G1 => G3 (present participles) had underlying G2 but surface G3. Thus, we had to constraint the dipth.simpl. for the surface representation only, hence "LowerG12" as the relevant variable, rather than "G12". Thus also ":j" etc. for all the digraph consonants.

Rule changes
============

Each rule is listed below, with comments added.

"Word Final Devoicing of Certain Single Consonants d3 etc.":
- formatting

"Word final weakening -tj and -ttj to -sj part 2":
- formatting

"Word Final Deletion of n2 m2 g2 h2":
- removed redundant 'matches' keyword

"oa:aÌ Diphtong Simplification Part I" and "... II":

This rule was the main target of this update, and all other changes are really side effects of the work on this rule. In order to facilitate the work, we numbered all the sub-rules, from 1 to 10 (each number 1...10 can be found both in Part I and Prt II, this of course due to the fact that twolc  cannot change more than one segment in one rule. We first comment a common change, then we comment the subrules one by one.

All subrules that should be blocked for diphthong simplification for G3, had to have the expression of their core consonant cluster changed. Earlier, the context at the second position of the right context said "Cns:+", this context was left only for the non-blocking rules. The blocking rules 3, 4, 6, 7, 9 got this string replaced with "LowerG12", in order to constrain the diphth.simpl to G12 only.

Subrules 1, 2, 5, 8
No changes.

Subrules 3, 4, 9
Only the Cns:+ to LowerG12 change (but see below for more comments on subrule 9).

Subrules 6, 7
Here, we rewrote the string of diacritics in the right context in order to make it comprehensive. This has to do with the deletion of the final subrule in the former version, in extenso "_ Vow     LowerG12  o:u ;". This rule was too tolerant, and had to be rewritten with all possible diacritics. These were imported into 6 and 7.

Subrule 9
This was the result of much grief and sorrow, but in the end we managed to fix the problems connected to it. As a matter of fact, this was where we found that we had to add the zeros to our G12 definitions.

Subrule 10
This rule got the same left context as the other. The "Vow" symbol directly to the right of "_" was outright wrong, and corrected to "a:". Finally, the regex in the far right of the context was rewritten., as it stood, it a.o. generated a string of two end-of-word symbols.

Finally, all old rules that had been lingering for a long time a comments were deleted

Here is a quote from our strategizing during the battle:
! Define diphtsimpl that occurs whenever we do not have G2 > G3 for Serie 2 nouns.

! Make sure dipht.simpl does not happen in derived G3 cases, even when the surface looks like G2
! <=> identify the Serie 2 G3 not as such, but via its causing context., ie endings or a marker
! Strategy: oa : aÌ <=> _ [G1] ...    [ Y6 | Y7 | Y9 ] ;
! Strategy: oa : aÌ <=> _      ...  - [ Y6 | Y7 | Y9 ] ; ! real G3, both overt G3 and G2 lookalikes

! Y6 = pres part/actor nouns, 2Du Imp
! Y7 = dim. of pres. part/actor nouns
! Y9 =  short passives of even-syllable verbs


The test suite got many new members, especially connected to the diphtong simplification issue.

"Ã¦:e Diphtong Simplification 1":
- formatting

"Consonant gradation b:0":
- formatting
- we added comments

"Consonant gradation d:0":
- formatting
- added comments

"Consonant gradation g:0":
"Consonant gradation k:0":
"Consonant gradation l:0":
"Consonant gradation m:0":
"Consonant gradation n:0":
"Consonant gradation p:0":
"Consonant gradation s:0":
"Consonant gradation nÌ:0":
"Consonant gradation f:0":
"Consonant gradation r:0":
"Consonant gradation v:0":
"Consonant gradation j:0":
"Consonant gradation t:0":
"Gradation Series 4, II-I, tj and ts":
- added comments

"bb:pp":
"gg:kk":
- added comments
- formatting

"dd:tt and dtj, dts":
- added comments

"Consonant insertion as II-III strengthening gradation with bm,gÅ":
- formatting


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/smj@4922 c7155fb1-f0a7-4240-a2fc-2600b6f42f90