Mostly non-essential changes, of three types:
- we made the file more readable, both the Sets, Definition and Rules section
- we corrected the examples (s0 to 0s, etc.)
- We added documentation text, to the sets and the definitions.

Then there were some substantial changes:

We corrected some errors in the Definition of LowerG2, which fixed one test case. This had mainly to do with the distribution of (Cns:0), as a result of such missing zeros we were prefiously not able to identify all G2.

We are now left with one test case that's not working:

! G2 > G3
!â¬ goarroY6X5jn
!â¬ goarru00jn
  #:0
  !
  â¬
  g
  o
  a
  r
  r
  o:u
  Y6:0
  X5:0
  j
  REJECTED: "oa:aÌ Diphtong Simplification Part I" fails in state 32.

This pair needs rethinking of the rule in question (and its complementary rule). The basic problem is the (non-)difference between G3 and G2 in pairs as the above.