Viken:^viken => Viken:Viken
The old version caused word forms with initial hyphen, which lead to a file named --init.plx, which again caused a lot of trouble when processing, as the filename was interpreted as an argument to the called program, and not as a file.

This error has already been fixed in sme.