Viken:^viken => Viken:Viken
The old version caused word forms with initial hyphen, which lead to a file named --init.plx, which again caused a lot of trouble when processing, as the filename was interpreted as an argument to the called program, and not as a file.

This error has already been fixed in sme.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/smj@11605 c7155fb1-f0a7-4240-a2fc-2600b6f42f90