The passive form was out-commented from intransitive verbs as it seems like most intransitive verbs do not take this form. Little research is done on the distribution of passives and of other derivations in general and this has to be looked at more thoroughly.
Added lexicons for 2 syllable loan words: Two lexicons for -it verbs (PRINTE_TV and TELTE_IV), and so far only one (trans) lexicon for -ut verbs (BRUKE_TV). Many of the words belonging in the latter lexicon are currently in STRÁFFUT lexicon which is itself a loan word. The name of this lexicon has to be changed.