added words from missing list
corrected existing entries:
endings -oga and -ija changed to -åvggå and -iddja
minor errors in Siri and Pippi
medijá words were previously corrected to medija, but the -ija ending is ungrammatical in julev. they were thus corrected back to medijá. -ia -> -ijá is as of yet not established as a general loan pattern rule, but at least it is grammatical according to Lule Sami grammar.