Merged two rules with identical insertion, and only slightly different contexts, in the hope that it will fix issues with PrsPrt and ImprtDual2 for VIEDJET_IV verbs in Hfst (Xerox has no problems). Unfortunately the change did not fix it, but I kept it since it makes the rules easier to read and debug.
Also removed superfluous, optional final right context for the merged rules and similar rules at the same place - a final, optional context doesn’t do anything: since it is optional and final, it is the same as specifying nothing.

Whitespace changes to make it easier to compare the rules and their context requirements.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/smj@110242 c7155fb1-f0a7-4240-a2fc-2600b6f42f90