A number of changes together with Sandra that resolved all remaining rule conflicts in both Xerox and Hfst. Some pair tests still fail Hfst, but the analyser behaves correctly, and they don’t fail in Xerox, so that might be a bug in hfst-pair-test instead.
There are a number of regressions in the yaml tests, but we’ll look at them separately - they are probably / hopefully caused by changed requirements on trigger symbols and other lexc contexts for the twolc rules, so that they can be fixed by changing the lexc code. Sandra will have a look at these.