Make lemma follow SE spelling
This is a word only used in Sweden, and it must therefore follow Swedish spelling rules for the lemma test to pass. We measure that we can generate the lemma (ie identity), and since this word is marked as SE only, the fst MUST generate the SE form. To get identity between generated and lemma, the lemma must also adhere to the SE spelling.