new lexicon TJIERREK-plc
Same as LONDON, but does not give Sem/Sur tag, since it is not usuall for SMJ place names to become surnames. The Sem/Sur tag made problems for grammar checker. Could not just remove the Sem/sur tag from LONDON-plc, or rename it to LONDON-plcsur since the lexicon is used by SMI